moves functions related to adding spin labels to their own file
